backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
et_xmlfile==2.0.0
fastapi==0.110.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.1.4
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.27.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
